+++ src/attribute_enum.hpp
@@ -51,7 +51,7 @@ namespace xios
             bool hasInheritedValue(void) const;                    
             
             bool isEqual(const CAttribute& attr );
-            int computeHash(void);
+            size_t computeHash(void);
 
             /// Destructeur ///
             virtual ~CAttributeEnum(void) { }
