+++ src/attribute_enum_impl.hpp
@@ -126,7 +126,7 @@ namespace xios
   }
 
   template <class T>
-  int CAttributeEnum<T>::computeHash(void)
+  size_t CAttributeEnum<T>::computeHash(void)
   {
     // Basic hash computation through string
     return std::hash<string>{}( toString() );
